complément view & liste de course

+++ database/migrations/2015_07_17_201030_create_lignesproduits_table.php
@@ -16,7 +16,7 @@ class CreateLignesproduitsTable extends Migration
 			$table->increments('id');
 			$table->integer('liste_id')->unsigned();
 			$table->integer('produit_id')->unsigned();
-			$table->integer('quantite');
+			$table->float('quantite');
 			$table->timestamps();
 
 			$table->foreign('liste_id')
